Customize what items you're trading!
Simply, for each line, add the item's market name (You can find that on backpack.tf, or the steam market, it's whatever name they use there) Then how much your selling an item IN REF. If you want to sell it in keys, great! use calculator.tf to see what that looks like (for example, 2.48 keys is 74.44 ref. And yes, the bot will accept keys for that). Then, write sell for selling, and buy for buying. Make sure to include the commas in-between!